+++ src/main/java/com/est/cranejob/comment/dto/request/UpdateCommentRequest.java
@@ -1,5 +1,6 @@
 package com.est.cranejob.comment.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.*;
@@ -17,7 +18,7 @@ import java.io.Serializable;
 @Builder
 public class UpdateCommentRequest implements Serializable {
 
-    @NotNull(message = "댓글 내용은 비어있을 수 없습니다.")
+    @NotBlank(message = "댓글 내용은 비어있을 수 없습니다.")
     @Size(max = 100, message = "댓글은 100자 이내로 작성 되어야 합니다.")
     private String content;
 }
