+++ src/main/java/com/est/cranejob/comment/dto/response/CommentDetailResponse.java
@@ -5,9 +5,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PastOrPresent;
 import jakarta.validation.constraints.Size;
-import lombok.Builder;
-import lombok.Data;
-import lombok.Value;
+import lombok.*;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -16,7 +14,10 @@ import java.time.LocalDateTime;
  * 댓글의 내용을 관리자만 볼 수 있는 형태로 응답하는 DTO<br>
  * DTO for {@link com.est.cranejob.comment.domain.Comment}
  */
-@Data
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
 @Builder
 public class CommentDetailResponse implements Serializable {
     private Long id;
